move only usage of withLibraryAttributeNamespaceCheck to calling service
